Add DelegatingCarrier
DelegatingCarrier is a carrier type which, for example, can be used
by protobufs (or any general struct that is part of an RPC payload
which already knows how to serialize itself) to copy the relevant
data inbetween itself and the Span.
